FFMiner version improvement ==> speed.
